requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
ecdsa==0.19.0
fastapi==0.115.4
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-jose==3.3.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
wheel==0.44.0
uvicorn
python-multipart